Extract config code from monolith main
