+++ generate_json_schema/src/main.rs
@@ -1,3 +1,6 @@
 fn main() {
-    println!("{}", dprint_plugin_typstyle::generate_json_schema());
+    println!(
+        "{}",
+        dprint_plugin_typstyle::configuration::generate_json_schema()
+    );
 }
